LIBITD-1975. Updated Person data mapping to match current mapping doc
Updated the "Person" object data mapping to match the mapping in the
"Online Staff Directory Mapping" Google Sheet in Google Drive.

Removed the "jobTitle" field from the "Person" object, as it is
no longer present in the mapping. Modified "descriptiveTitle" to
be retrieved from the LDAP "umDisplayTitle" to match the mapping.

Added a "docs/StaffDirectory.md" file to document the mapping
from source documents to fields on the "Person" object.

https://issues.umd.edu/browse/LIBITD-1975