LIBWEB-5367. Modified code to retrieve field mappings from Google doc
Modified the code to retrieve the fields mappings from a
"Field Mappings" sheet in the Google Sheets document.

This enables the Google Sheets document to configure the matching of
the fields from a source ("Staff" sheet, "Organization" sheets, or LDAP)
with a field in the destination.

The "Person" class then just becomes a data object for holding the
source fields about a person. The "ExcelGenerator" class then uses
the FieldMappings map to determine the "source" and "sourceField" to
retrieve from the Person object.

"Derived" fields are still generated manually within the
"ExcelGenerator" class.

https://issues.umd.edu/browse/LIBWEB-5367