LIBITD-1975. Added handling for "Display Type" column
Added handling for "Percentage" and "CategoryStatus" DisplayType fields.

For "CategoryStatus", read in the "CategoryStatus" sheet containing the
mapping for abbreviations to full text, and used it to populate the
spreadsheet.

Converted "ExcelGenerator.generate" from a static method, to an
instance method (initializing the fieldMappings and
categoryStatusAbbreviations in the constructor) to simplify passing
information around the various method.

https://issues.umd.edu/browse/LIBITD-1975