LIBITD-1975. Split processing into StaffRetriever/AllStaffListBuilder
Split processing to a "StaffRetriever" class, which retrieves the
staff data from the Google Sheets document and parses it into JSON,
and an "AllStaffListBuilder" class that parses the JSON and converts
it into an Excel spreadsheet.

This will enable the JSON file to be used by multiple "builder" classes,
such as a future one for Drupal. It also provides a mechanism for
determining if the staff data has changed by comparing it to an
existing JSON file (this comparison functionality has not yet been
implemented).

Modified the Maven configuration to use the "Application Assembler"
plugin to generate "staff-retriever" and "all-staff-list-builder"
scripts, as well as import the Jackson JSON library.

Removed some unnecessary "slf4j" libraries that were causing warnings
when running Maven.

https://issues.umd.edu/browse/LIBITD-1975