LIBITD-1975. Modified StaffRetriever to sort persons by last/first name
Added "PersonSorter" class to sort Persons based on last name and
first name (based on LDAP attributes). Needed to have a separate
class to enable sorting on both last and first names (instead of
just last names), because last names are common, and would end up
in a different order than the order in the AllStaffList.xlsx
spreadsheet.

https://issues.umd.edu/browse/LIBITD-1975