LIBITD-1975. Refactored StaffRetriever, putting more logic into classes
Refactored StaffRetriever main class, putting more of the logic into
the "Organization" and "Staff" classes. Renamed "Organizations" class
to "Organization" to match the name of the sheet in the Google Sheets
document.

Added "Staff" class to parallel the "Organization" class.

Moved sorter for Persons into the "Persons" class as a static class and
renamed to "LastNameFirstNameComparator" for clarity.

https://issues.umd.edu/browse/LIBITD-1975