LIBITD-1976. Modified Excel spreadsheet to set "ReadOnly-Recommended"
Modified the Excel spreadsheet to set the "ReadOnly-Recommended" flag
to true. This requires a user and password, so an "excelUser" parameter
to the "ExcelGenerator" class, and am (optional)  "--user" flag was
added to the "AllStaffListBuilder" command-line arguments,

Setting the Excel spreadsheet to "ReadOnly-Recommended" is necessary for
the CronJob, as opening the file for modification prevents the CronJob
from updating the file on the Samba share.

https://issues.umd.edu/browse/LIBITD-1976